# LibWeb: Make named character reference tokenization more spec-compliant & efficient