--- a/Libraries/LibWeb/CMakeLists.txt
+++ b/Libraries/LibWeb/CMakeLists.txt
@@ -918,6 +918,8 @@ invoke_generator(
 
 generate_css_implementation()
 
+generate_html_implementation()
+
 invoke_generator(
     "WebGLRenderingContextImpl.cpp"
     Lagom::GenerateWebGLRenderingContext
@@ -961,6 +963,7 @@ set(GENERATED_SOURCES
     WebGL/WebGLRenderingContextImpl.cpp
     Worker/WebWorkerClientEndpoint.h
     Worker/WebWorkerServerEndpoint.h
+    HTML/Parser/NamedCharacterReferences.cpp
 )
 
 serenity_lib(LibWeb web)