--- a/Libraries/LibWeb/HTML/Parser/HTMLTokenizer.h
+++ b/Libraries/LibWeb/HTML/Parser/HTMLTokenizer.h
@@ -14,6 +14,7 @@
 #include <AK/Utf8View.h>
 #include <LibGC/Ptr.h>
 #include <LibWeb/Forward.h>
+#include <LibWeb/HTML/Parser/Entities.h>
 #include <LibWeb/HTML/Parser/HTMLToken.h>
 
 namespace Web::HTML {
@@ -210,6 +211,8 @@ private:
 
     HTMLToken m_current_token;
     StringBuilder m_current_builder;
+
+    NamedCharacterReferenceMatcher m_named_character_reference_matcher;
 
     Optional<ByteString> m_last_emitted_start_tag_name;