--- a/Libraries/LibWeb/HTML/Parser/HTMLTokenizerHelpers.cpp
+++ b/Libraries/LibWeb/HTML/Parser/HTMLTokenizerHelpers.cpp
@@ -5,7 +5,9 @@
  */
 
 #include <LibTextCodec/Decoder.h>
+#include <LibWeb/HTML/Parser/Entities.h>
 #include <LibWeb/HTML/Parser/HTMLTokenizerHelpers.h>
+#include <LibWeb/HTML/Parser/NamedCharacterReferences.h>
 
 namespace Web::HTML {
 
@@ -22,9 +24,27 @@ OptionalString decode_to_utf8(StringView text, StringView encoding)
 
 OptionalEntityMatch match_entity_for_named_character_reference(StringView entity)
 {
-    auto entity_match = code_points_from_entity(entity);
-    if (entity_match.has_value())
-        return entity_match.release_value();
+    NamedCharacterReferenceMatcher matcher;
+    int consumed_length = 0;
+    for (auto c : entity) {
+        if (!matcher.try_consume_ascii_char(c))
+            break;
+        consumed_length++;
+    }
+
+    auto codepoints = matcher.code_points();
+    if (codepoints.has_value()) {
+        EntityMatch match;
+        auto matched_length = consumed_length - matcher.overconsumed_code_points();
+        auto matched_string_view = entity.substring_view(0, matched_length);
+        auto second_codepoint = named_character_reference_second_codepoint_value(codepoints.value().second);
+        if (second_codepoint.has_value()) {
+            match = { { codepoints.value().first, second_codepoint.release_value() }, matched_string_view };
+        } else {
+            match = { { codepoints.value().first }, matched_string_view };
+        }
+        return match;
+    }
     return std::nullopt;
 }