… Libraries/LibWeb/HTML/Parser/Entities.cpp
… Libraries/LibWeb/HTML/Parser/Entities.json
… Meta/Lagom/Tools/CodeGenerators/LibWeb/GenerateNamedCharacterReferences.cpp