This PR focuses on named character reference tokenization, both in terms of correctness and performance.

Tons of details in the commit message, so see that for the full explanation. This implementation is also based on [this experiment of mine using the same idea written in Zig](https://github.com/squeek502/named-character-references), and there are more nitty-gritty details available in its README if you're interested.

The tl;dr is that a [DAFSA](https://en.wikipedia.org/wiki/Deterministic_acyclic_finite_state_automaton) is generated at compile-time and used for efficient codepoint-by-codepoint named character reference matching.

I was not able to get a working build with `-DENABLE_SWIFT` (instructions on this would be appreciated), so I left the Swift HTMLTokenizer implementation alone and used the new NamedCharacterReferenceMatcher to satisfy the previous `match_entity_for_named_character_reference` API. In other words, the Swift version will work exactly the same as it did before this PR.

If it's helpful for reviewing, here's what the generated NamedCharacterReferences.h/cpp look like:
https://gist.github.com/squeek502/550164a8d5549e8a2bf45ece4ce38045

---

In a benchmark using [an arbitrary set of test files found online](https://github.com/AndreasMadsen/htmlparser-benchmark/tree/master/files), the tokenizer is about 1.23x faster (2070ms to 1682ms).

<details>
<summary>General benchmark code</summary>

```c++
BENCHMARK_CASE(benchfiles)
{
    MUST(Core::Directory::for_each_entry("files"sv, Core::DirIterator::SkipParentAndBaseDir, [&](Core::DirectoryEntry const& entry, Core::Directory const& parent) {
        auto file = MUST(parent.open(entry.name, Core::File::OpenMode::Read));
        auto file_size = MUST(file->size());
        auto content = MUST(ByteBuffer::create_uninitialized(file_size));
        MUST(file->read_until_filled(content.bytes()));
        ByteString file_contents { content.bytes() };
        auto tokens = run_tokenizer(file_contents);
        return IterationDecision::Continue;
    }));
}
```

</details>

(this was a bit surprising, I wasn't expecting this to affect the average case much)

---

In a benchmark hyper-focused on named character references (see [`named-char-test.html` here](https://gist.github.com/squeek502/07b7dee1086f6e9dc38c4a880addfeca) for the test file used, it's just tens of thousands of valid and invalid named character references), the tokenizer is ~8x faster (758ms to 93ms).

<details>
<summary>Named character reference benchmark code</summary>

```c++
BENCHMARK_CASE(named_character_references)
{
    StringView path = "named-char-test.html"sv;

    auto file = MUST(Core::File::open(path, Core::File::OpenMode::Read));
    auto file_size = MUST(file->size());
    auto content = MUST(ByteBuffer::create_uninitialized(file_size));
    MUST(file->read_until_filled(content.bytes()));
    ByteString file_contents { content.bytes() };
    auto tokens = run_tokenizer(file_contents);
    dbgln("{} tokens", tokens.size());
}
```

</details>

---

The amount of data stored for the named character reference matching has also been reduced by around 95KiB (just going off the size of `liblagom-web.so` before/after these changes).

---

With this commit, many WPT subtests now pass, and when combined together with https://github.com/LadybirdBrowser/ladybird/pull/3163, AFAICT all character reference-related subtests are now passing. Here are the results from running the WPT tests in `/html/syntax/parsing` (comparing master to this branch + https://github.com/LadybirdBrowser/ladybird/pull/3163):

<details>
<summary>WPT results for /html/syntax/parsing</summary>

```
  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities01.html?run_type=write:
  └ PASS [expected FAIL] html5lib_entities01.html 6b336a43e394d3ab7ceb2ab54c63409e8a27aded

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities01.html?run_type=write:
  └ PASS [expected FAIL] html5lib_entities01.html d21511e2df56c306c78e1449c960c66e565e016e

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities01.html?run_type=write:
  └ PASS [expected FAIL] html5lib_entities01.html 39107d16f24d4c7bcd40ad1239b5f4f677877ee8

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities01.html?run_type=write:
  └ PASS [expected FAIL] html5lib_entities01.html bffe7b00046407080251ab6bf58cb97ce2a34893

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities01.html?run_type=write:
  └ PASS [expected FAIL] html5lib_entities01.html 5aef37f1f2b9ac45adfade044c882eb09a297569

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities01.html?run_type=write:
  └ PASS [expected FAIL] html5lib_entities01.html 6e2d817539fb3b2023c7bcb88ad220c136f70cf0

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities01.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_entities01.html 16c694bcf0b3ff3723fa070eea7e1e82ef12a337

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities01.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_entities01.html 05e04b39ef06e2367a33326f5dd566913aa6628f

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities01.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_entities01.html fbf7d9fec595585869c5c595d5588b34fd175278

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities01.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_entities01.html e59b0a76d7bcfb429b27e00e469f35e08a9bdd1a

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities01.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_entities01.html 5ea854d6ecd4d6dd459cb36d4faf3ed36e11c073

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities01.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_entities01.html 119cd15b852615cd0fce759769b4a3788595e3bb

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities01.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_entities01.html bf6c90305b2856c2d9c9a146dfff867fe7a5e0f3

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities01.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_entities01.html 6b336a43e394d3ab7ceb2ab54c63409e8a27aded

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities01.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_entities01.html d21511e2df56c306c78e1449c960c66e565e016e

  ▶ Unexpected subtest result in /html/syntsingle:
  └ PASS [expected FAIL] html5lib_entities01.html 39107d16f24d4c7bcd40ad1239b5f4f677877ee8

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities01.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_entities01.html bffe7b00046407080251ab6bf58cb97ce2a34893

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities01.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_entities01.html 5aef37f1f2b9ac45adfade044c882eb09a297569

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities01.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_entities01.html 6e2d817539fb3b2023c7bcb88ad220c136f70cf0

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities02.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_entities02.html ea66863900b0b42deee5a77c58a432c2215c32ac

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities02.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_entities02.html fe22904d5f3936bedc1fa110e6bde48895b399a0

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities02.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_entities02.html 6553483a30141fcff05787287c2c212df9f468e8

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities02.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_entities02.html 88d7c74afcb27bbee3e3255d9116dce9c3dc6d73

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities02.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_entities02.html db5d22d3350e0a51d675dc17c641c73251a4739d

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities02.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_entities02.html ea08276faa7ba526e612fc1e80047d705cd29885

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities02.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_entities02.html f9d3950620f8adcbe5f9a0542c7967de4be65963

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities02.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_entities02.html 8e35dacd7c296f054e58f1ce83719401c8aff8a0

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities02.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_entities02.html 565c5f6744a27602bb466d6df77803a80f064752

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities02.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_entities02.html f908b529ac9ca5366e1160856db2c3d17e3898c9

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities02.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_entities02.html 1294ffc6bee2ee41f65a60ac48ba445b99504286

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities02.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_entities02.html ba7d8cdd4b40020f7af6bdde75a3574b5771fac9

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities02.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_entities02.html 6bbeec30b849cebd1366ebb2e6d2a6c1790e8c68

  ▶ OK [expected TIMEOUT] /html/syntax/parsing/html5lib_entities01.html?run_type=uri

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities01.html?run_type=uri:
  └ PASS [expected FAIL] html5lib_entities01.html 6b336a43e394d3ab7ceb2ab54c63409e8a27aded

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities01.html?run_type=uri:
  └ PASS [expected FAIL] html5lib_entities01.html d21511e2df56c306c78e1449c960c66e565e016e

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities01.html?run_type=uri:
  └ PASS [expected FAIL] html5lib_entities01.html 39107d16f24d4c7bcd40ad1239b5f4f677877ee8

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities01.html?run_type=uri:
  └ PASS [expected TIMEOUT] html5lib_entities01.html ceba8404405dd3b3b423c45411bde15bf72a846d

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities01.html?run_type=uri:
  └ PASS [expected FAIL] html5lib_entities01.html bffe7b00046407080251ab6bf58cb97ce2a34893

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities01.html?run_type=uri:
  └ PASS [expected FAIL] html5lib_entities01.html 5aef37f1f2b9ac45adfade044c882eb09a297569

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities01.html?run_type=uri:
  └ PASS [expected FAIL] html5lib_entities01.html 6e2d817539fb3b2023c7bcb88ad220c136f70cf0

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_entities01.html?run_type=uri:
  └ PASS [expected TIMEOUT] html5lib_entities01.html d2584faaa4dda5283955b2dc22812a018d04a72d

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_html5test-com.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_html5test-com.html be72b058e5be0f6aef2c442d83c92c0d251fcb7f

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_html5test-com.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_html5test-com.html ab6e31cf52c8d57d6dfdcaf7165f1abf7bd5e73d

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_html5test-com.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_html5test-com.html 11240d9b03b14eb515d6a1d1595c5a409830ea38

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_html5test-com.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_html5test-com.html 809c1bebcded8f43981af902442ff8a2db5d2578

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_html5test-com.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_html5test-com.html bcbeb84f40e56a642b794d514e97e3ec303d4a79

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_tests16.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_tests16.html 1553cebdf01dc953ed7983d39a18752a4fbb24d7

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_tests16.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_tests16.html fea28aab54637701c5dfaef4f3fe64c72b272e1c

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_tests2.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_tests2.html e0c43080cf61c0696031bdb097bea4f2a647cfc2

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_tests2.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_tests2.html 8dc47e70b94f2bea514ceaa51153ec1beeeda7ef

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_tests2.html?run_type=write:
  └ PASS [expected FAIL] html5lib_tests2.html e0c43080cf61c0696031bdb097bea4f2a647cfc2

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_tests2.html?run_type=uri:
  └ PASS [expected FAIL] html5lib_tests2.html e0c43080cf61c0696031bdb097bea4f2a647cfc2

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_tests24.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_tests24.html 692c2dbacf18cb758f26a3d9e7d9add4356f9067

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_tests24.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_tests24.html 614cc9827d3a0c5f91863dde5281dd0d97f64e6d

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_tests24.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_tests24.html c517924583ee71b8e684c9ca1f2eed5e88139a39

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_tests24.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_tests24.html 140a82fab878c139b388d159c511eb999fe2d8c7

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_tests24.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_tests24.html a2ddddcccb652a6529daafd4153a0e12b6d5ca8c

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_tests24.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_tests24.html da2e30a0b6577b608bf48bbd11a16ff832bc7e46

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_tests24.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_tests24.html 66a5777f5453bd4b5161f00df02883b6d71f7cea

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_tests24.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_tests24.html c8d97f31b70f67005eeacc3c86ac29e577c3d0ed

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_tests5.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_tests5.html bd7dfd1a0f74731c22b3e2d331f7c14ba7c9a4e8

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_tests5.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_tests5.html 5f847a390a413a42fcef3d4510ddc56815c7d722

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_tests6.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_tests6.html fa05c524ac7918197adf422a2c4be35d5eca9ddc

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_tests6.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_tests6.html 652ae0a8ca3ab725b2d10e9866898b3419333f64

  ▶ Unexpected subtest result in /html/syntax/parsing/html5lib_webkit01.html?run_type=write_single:
  └ PASS [expected FAIL] html5lib_webkit01.html 3bea2bf663be5de2bbcdad57ac95c5933e266d42

  ▶ Unexpected subtest result in /html/syntax/parsing/ambiguous-ampersand.html:
  └ PASS [expected FAIL] Check div structure: document.write

Ran 209 tests finished in 111.6 seconds.
  • 195 ran as expected. 0 tests skipped.
  • 1 tests unexpectedly okay
  • 14 tests had unexpected subtest results
```

</details>

There are probably other newly passing tests/subtests elsewhere (one I'm aware of is `html/webappapis/dynamic-markup-insertion/document-write/041.html`)