@ADKaster 2025-01-06
@squeek502 I mentioned this on discord, but just to restate here: The last commit needs some extra review, but the commits before that that strictly fix bugs without refactoring are non-controversial. If you split out the last commit into its own PR I'll merge the precursors.