@squeek502 2025-01-06
Makes sense; split the straightforward bug fixes into https://github.com/LadybirdBrowser/ladybird/pull/3163