@squeek502 2025-03-17
Thanks for the review, will address those comments soon.

> I'm also mildly curious if building this as a dynamic trie is much slower

It wouldn't be slower, but would need more memory to store the data. The goal here is to be just as fast as a full trie, but using a compact representation of the data.

Here's a pull request for a Zig HTML parser library that compares a (generated) trie to the DAFSA strategy used in this PR: https://github.com/chadwain/rem/pull/15 (after I made some changes it performs roughly the same, see https://github.com/chadwain/rem/pull/15#issuecomment-2178394880 [and those changes are included in the Ladybird version]).